Moved get_*_for to lib/util
get_reader_for, get_writer_for and get_parser_for now in lib/util

MMTFReader and MMTFParser now accept a MMTFDecoder object as 'filename'
