+++ package/MDAnalysis/topology/core.py
@@ -41,35 +41,7 @@ from .guessers import (
     get_atom_mass, guess_atom_mass, guess_atom_charge,
     guess_bonds, guess_angles, guess_dihedrals, guess_improper_dihedrals,
 )
-from ..lib.util import cached
+from ..lib.util import cached, get_parser_for
 from ..lib import util
 
-
-def get_parser_for(filename, format=None):
-    """Return the appropriate topology parser for *filename*.
-
-    Automatic detection is disabled when an explicit *format* is
-    provided.
-
-    :Raises:
-      *ValueError*
-        If no appropriate parser could be found.
-    """
-    if format is None:
-        format = util.guess_format(filename)
-    format = format.upper()
-    try:
-        return _PARSERS[format]
-    except KeyError:
-        raise ValueError(
-            "Cannot autodetect topology type for file '{0}' "
-            "(file extension could not be parsed).\n"
-            "           You can use 'Universe(topology, ..., topology_format=FORMAT)' "
-            "to explicitly specify the format and\n"
-            "           override automatic detection. Known FORMATs are:\n"
-            "           {1}\n"
-            "           See http://docs.mdanalysis.org/documentation_pages/topology/init.html#supported-topology-formats\n"
-            "           For missing formats, raise an issue at "
-            "http://issues.mdanalysis.org".format(filename, _PARSERS.keys()))
-
-
+#tumbleweed
