+++ testsuite/MDAnalysisTests/topology/test_mmtf.py
@@ -3,6 +3,7 @@ from numpy.testing import (
     assert_,
     assert_array_equal,
 )
+import mmtf
 
 import MDAnalysis as mda
 from MDAnalysis.core.groups import AtomGroup
@@ -72,6 +73,12 @@ class TestMMTFUniverse(object):
         assert_(all(self.u.atoms.models == 0))
 
 
+class TestMMTFUniverseFromDecoder(TestMMTFUniverse):
+    def setUp(self):
+        top = mmtf.parse(MMTF)
+        self.u = mda.Universe(top)
+
+
 class TestMMTFgzUniverse(object):
     def setUp(self):
         self.u = mda.Universe(MMTF_gz)
@@ -98,3 +105,7 @@ class TestMMTFgzUniverse(object):
         assert_(len(m1) == 570)
         assert_(len(m2) == 570)
 
+class TestMMTFgzUniverseFromDecoder(TestMMTFgzUniverse):
+    def setUp(self):
+        top = mmtf.parse_gzip(MMTF_gz)
+        self.u = mda.Universe(top)
